Fix a GCC warning in several vector and matrix class tests

+++ blazetest/src/mathtest/compressedmatrix/ClassTest1.cpp
@@ -1134,11 +1134,11 @@ void ClassTest::testAssignment()
    {
       test_ = "Row-major/row-major CompressedMatrix sparse matrix assignment stress test";
 
-      typedef blaze::CompressedMatrix<unsigned int,blaze::rowMajor>  RandomMatrixType;
+      typedef blaze::CompressedMatrix<short,blaze::rowMajor>  RandomMatrixType;
 
       blaze::CompressedMatrix<int,blaze::rowMajor> mat1;
-      const unsigned int min( randmin );
-      const unsigned int max( randmax );
+      const short min( randmin );
+      const short max( randmax );
 
       for( size_t i=0UL; i<100UL; ++i )
       {
@@ -1193,11 +1193,11 @@ void ClassTest::testAssignment()
    {
       test_ = "Row-major/column-major CompressedMatrix sparse matrix assignment stress test";
 
-      typedef blaze::CompressedMatrix<unsigned int,blaze::columnMajor>  RandomMatrixType;
+      typedef blaze::CompressedMatrix<short,blaze::columnMajor>  RandomMatrixType;
 
       blaze::CompressedMatrix<int,blaze::rowMajor> mat1;
-      const unsigned int min( randmin );
-      const unsigned int max( randmax );
+      const short min( randmin );
+      const short max( randmax );
 
       for( size_t i=0UL; i<100UL; ++i )
       {
@@ -1747,11 +1747,11 @@ void ClassTest::testAssignment()
    {
       test_ = "Column-major/row-major CompressedMatrix sparse matrix assignment stress test";
 
-      typedef blaze::CompressedMatrix<unsigned int,blaze::rowMajor>  RandomMatrixType;
+      typedef blaze::CompressedMatrix<short,blaze::rowMajor>  RandomMatrixType;
 
       blaze::CompressedMatrix<int,blaze::columnMajor> mat1;
-      const unsigned int min( randmin );
-      const unsigned int max( randmax );
+      const short min( randmin );
+      const short max( randmax );
 
       for( size_t i=0UL; i<100UL; ++i )
       {
@@ -1807,11 +1807,11 @@ void ClassTest::testAssignment()
    {
       test_ = "Column-major/column-major CompressedMatrix sparse matrix assignment stress test";
 
-      typedef blaze::CompressedMatrix<unsigned int,blaze::columnMajor>  RandomMatrixType;
+      typedef blaze::CompressedMatrix<short,blaze::columnMajor>  RandomMatrixType;
 
       blaze::CompressedMatrix<int,blaze::columnMajor> mat1;
-      const unsigned int min( randmin );
-      const unsigned int max( randmax );
+      const short min( randmin );
+      const short max( randmax );
 
       for( size_t i=0UL; i<100UL; ++i )
       {
